rag_rhyme/requirements.txt
faiss-cpu
transformers
datasets
sentence-transformers
scikit-learn
pandas
numpy